// types/ticket.ts
export type Ticket = {
  id: string;
  ticketid: string;
  adminreplying: string;
  userid: string;
  contactid: string;
  email: string;
  name: string;
  department: string;
  priority: string;
  status: string;
  service: string;
  ticketkey: string;
  subject: string;
  message: string;
  admin: string;
  date: string;
  project_id: string;
  lastreply: string;
  clientread: string;
  adminread: string;
  assigned: string;
  line_manager: string;
  milestone: string;
}